Rename Python package oif to openinterfaces to have a correct full package name
