--- lang_python/oif_impl/oif/_impl/ivp/scipy_ode/scipy_ode.conf
@@ -1,2 +1,0 @@
-python
-oif._impl.ivp.scipy_ode.scipy_ode ScipyODE
