--- lang_python/oif_impl/oif/_impl/linsolve/numpy/numpy.conf
@@ -1,2 +1,0 @@
-python
-oif._impl.linsolve.numpy.linsolve NumPyLinsolve
