--- lang_python/oif_impl/oif/_impl/qeq/py_qeq_solver/py_qeq_solver.conf
@@ -1,2 +1,0 @@
-python
-oif._impl.qeq.py_qeq_solver.qeq_solver QeqSolver
