+++ lang_python/oif_impl/openinterfaces/_impl/ivp/scipy_ode/scipy_ode.conf
@@ -1,0 +1,2 @@
+python
+openinterfaces._impl.ivp.scipy_ode.scipy_ode ScipyODE
