+++ lang_python/oif_impl/openinterfaces/_impl/linsolve/numpy/numpy.conf
@@ -1,0 +1,2 @@
+python
+openinterfaces._impl.linsolve.numpy.linsolve NumPyLinsolve
