+++ lang_python/oif_impl/openinterfaces/_impl/qeq/py_qeq_solver/py_qeq_solver.conf
@@ -1,0 +1,2 @@
+python
+openinterfaces._impl.qeq.py_qeq_solver.qeq_solver QeqSolver
