+++ lang_python/oif_impl/openinterfaces/_impl/qeq/py_qeq_solver/qeq_solver.py
@@ -1,7 +1,7 @@
 import math
 
 import numpy as np
-from oif._impl.interfaces.qeq import QeqInterface
+from openinterfaces._impl.interfaces.qeq import QeqInterface
 
 
 class QeqSolver(QeqInterface):
