+++ lang_python/oif_interfaces/openinterfaces/interfaces/linear_solver.py
@@ -9,7 +9,7 @@ where :math:`A` is a square matrix and :math:`b` is a vector.
 """
 
 import numpy as np
-from oif.core import OIFPyBinding, load_impl, unload_impl
+from openinterfaces.core import OIFPyBinding, load_impl, unload_impl
 
 
 class LinearSolver:
